main makefile refactor
