+++ srcs/parsing/prs_commandline.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parse_commandline.c                                :+:      :+:    :+:   */
+/*   prs_commandline.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: znichola <znichola@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/12 21:13:42 by znichola          #+#    #+#             */
-/*   Updated: 2023/01/13 18:38:31 by znichola         ###   ########.fr       */
+/*   Updated: 2023/01/14 09:04:49 by znichola         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
