+++ srcs/parsing/prs_pipeline.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parse_pipeline.c                                   :+:      :+:    :+:   */
+/*   pipeline.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: znichola <znichola@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/13 08:14:59 by znichola          #+#    #+#             */
-/*   Updated: 2023/01/13 18:39:09 by znichola         ###   ########.fr       */
+/*   Updated: 2023/01/14 09:03:37 by znichola         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
